+++ lib/cache.js
@@ -5,7 +5,7 @@
 const Redis = require('ioredis');
 const { isNil } = require('lodash');
 
-class Cache {
+module.exports = class Cache {
   constructor({ disabled = false, redisUri }) {
     this.disabled = disabled;
     this.redisUri = redisUri;
@@ -84,6 +84,4 @@ class Cache {
     await this.redis.set(Cache.getKey(key), JSON.stringify({ type, value }));
     await this.redis.expire(key, ttl);
   }
-}
-
-module.exports = Cache;
+};
